+++ hasklepias-core/src/Cohort/Core.hs
@@ -65,6 +65,7 @@ import           Data.Map.Strict               as Map
                                                 ( Map )
 import           Data.Maybe                     ( Maybe(..)
                                                 , mapMaybe
+                                                , catMaybes
                                                 )
 import           Data.Monoid                    ( mempty )
 import           Data.Ord                       ( Ord(..) )
@@ -170,14 +171,14 @@ getSubjectData (MkSubject (_, x)) = x
 -- into the desired return type.
 data CohortSpec d1 d0 i a = MkCohortSpec
   { runIndices  :: d1 -> IndexSet i a
-  , runCriteria :: d1 -> Criteria
+  , runCriteria :: Index i a -> d1 -> Criteria
   , runFeatures :: Index i a -> d1 -> d0
   }
 
 -- | Creates a @'CohortSpec'@.
 specifyCohort
   :: (d1 -> IndexSet i a)
-  -> (d1 -> Criteria)
+  -> (Index i a -> d1 -> Criteria)
   -> (Index i a -> d1 -> d0)
   -> CohortSpec d1 d0 i a
 specifyCohort = MkCohortSpec
@@ -188,38 +189,46 @@ evalIndicesOnSubjects
 evalIndicesOnSubjects spec (MkPopulation pop) =
   fmap (\x -> (runIndices spec (getSubjectData x), x)) pop
 
-data Step1 d i a = MkStep1 (ObsUnit d) (Maybe (Index i a))
+data Step1 d i a = MkStep1 (ObsUnit d) (Index i a)
 
-doStep1 :: (IndexSet i a, Subject d1) -> [Step1 d1 i a]
+doStep1 :: (IndexSet i a, Subject d1) -> [Maybe (Step1 d1 i a)]
 doStep1 (MkIndexSet indices, MkSubject (id, d)) = case indices of
-  Nothing  -> pure $ MkStep1 (MkObsUnit (makeObsID 1 id) d) Nothing
-  Just set -> zipWith
-    (\x y -> MkStep1 (MkObsUnit (makeObsID y id) d) (Just x))
+  Nothing  -> pure Nothing
+  Just set ->
+    zipWith
+    (\x y -> Just $ MkStep1 (MkObsUnit (makeObsID y id) d) x)
     (Set.toList set)
     [1 ..]
 
+  -- Nothing  -> pure $ MkStep1 (MkObsUnit (makeObsID 1 id) d) Nothing
+  -- Just set -> zipWith
+  --   (\x y -> MkStep1 (MkObsUnit (makeObsID y id) d) (Just x))
+  --   (Set.toList set)
+  --   [1 ..]
+
 -- | Evaluates the @'runIndices'@ of a @'CohortSpec'@ on a @'Population'@ to 
 -- return a list of @(Index i a, ObsUnit d)@ (one per observational unit). 
-evalIndices :: CohortSpec d1 d0 i a -> Population d1 -> [Step1 d1 i a]
+evalIndices :: CohortSpec d1 d0 i a -> Population d1 -> [Maybe (Step1 d1 i a)]
 evalIndices spec x = doStep1 =<< evalIndicesOnSubjects spec x
 
 newtype Step2 d i a = MkStep2 (Criteria, Step1 d i a)
 
 -- | Evaluates the @'runCriteria'@ of a @'CohortSpec'@ on a @'Population'@ to 
 -- return a list of @(Criteria, Index i a, ObsUnit d1)@. 
-evalCriteria :: CohortSpec d1 d0 i a -> [Step1 d1 i a] -> [Step2 d1 i a]
-evalCriteria spec = fmap
-  (\(MkStep1 ou mi) -> MkStep2 (runCriteria spec (obsData ou), MkStep1 ou mi))
+evalCriteria :: CohortSpec d1 d0 i a -> [Maybe (Step1 d1 i a)] -> [Maybe (Step2 d1 i a)]
+evalCriteria spec = fmap (fmap
+  (\(MkStep1 ou i) -> MkStep2 (runCriteria spec i (obsData ou), MkStep1 ou i)))
 
 data Step3 d i a =
     I1 CohortStatus
   | I2 CohortStatus (ObsUnit d) (Index i a)
 
-doStep3 :: Step2 d i a -> Step3 d i a
-doStep3 (MkStep2 (c, MkStep1 ou i)) = case (status, i) of
-  (Included, Just index) -> I2 status ou index
-  _                      -> I1 status
-  where status = checkCohortStatus i c
+doStep3 :: Maybe (Step2 d i a) -> Step3 d i a
+doStep3 (Just (MkStep2 (c, MkStep1 ou i)))= case (status, i) of
+  (Included, index) -> I2 status ou index
+  _                 -> I1 status
+  where status = checkCohortStatus (Just i) c
+doStep3 Nothing = I1 SubjectHasNoIndex 
 
 step3toStatus :: Step3 d i a -> CohortStatus
 step3toStatus (I1 x    ) = x
@@ -227,7 +236,7 @@ step3toStatus (I2 x _ _) = x
 
 -- | Convert a list of @(Criteria, Index i a, ObsUnit d)@ into a list of 
 --   @(CohortStatus, Index i a, ObsUnit d)@
-evalCohortStatus :: [Step2 d i a] -> [Step3 d i a]
+evalCohortStatus :: [Maybe (Step2 d i a)] -> [Step3 d i a]
 evalCohortStatus = fmap doStep3
 
 -- | Internal function for mapping an @ObsUnit d1@ along with its corresponding 
@@ -250,7 +259,7 @@ evalCohort spec pop = MkCohort
   indices         = evalIndices spec pop
   unitsWithCrits  = evalCriteria spec indices
   unitsWithStatus = evalCohortStatus unitsWithCrits
-  firstCrit       = (\(MkStep2 (c, _)) -> c) <$> headMay unitsWithCrits
+  firstCrit       = (\(MkStep2 (c, _)) -> c) <$> headMay (catMaybes unitsWithCrits)
   statuses        = step3toStatus <$> unitsWithStatus
 
 {-| A container hold multiple cohorts of the same type. The key is the name of 
@@ -268,7 +277,7 @@ newtype CohortSetSpec d1 d0 i a = MkCohortSetSpec (Map Text (CohortSpec d1 d0 i 
 
 -- | Make a set of 'CohortSpec's from list input.
 makeCohortSpecs
-  :: [(Text, d1 -> IndexSet i a, d1 -> Criteria, Index i a -> d1 -> d0)]
+  :: [(Text, d1 -> IndexSet i a, Index i a -> d1 -> Criteria, Index i a -> d1 -> d0)]
   -> CohortSetSpec d1 d0 i a
 makeCohortSpecs l = MkCohortSetSpec
   $ fromList (fmap (\(n, i, c, f) -> (n, specifyCohort i c f)) l)